Rename using React-native rename
