+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.guff",
+    package = "com.practica",
 )
 
 android_resource(
     name = "res",
-    package = "com.guff",
+    package = "com.practica",
     res = "src/main/res",
 )
