+++ android/settings.gradle
@@ -1,3 +1,3 @@
-rootProject.name = 'guff'
+rootProject.name = 'Practica'
 
 include ':app'
