+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "guff",
-  "displayName": "guff"
+  "name": "Practica",
+  "displayName": "Practica"
 }
