+++ ios/Practica/Images.xcassets/AppIcon.appiconset/Contents.json
