+++ ios/PracticaTests/Info.plist
