+++ ios/PracticaTests/PracticaTests.m
@@ -16,11 +16,11 @@
 #define TIMEOUT_SECONDS 600
 #define TEXT_TO_LOOK_FOR @"Welcome to React Native!"
 
-@interface guffTests : XCTestCase
+@interface PracticaTests : XCTestCase
 
 @end
 
-@implementation guffTests
+@implementation PracticaTests
 
 - (BOOL)findSubviewInView:(UIView *)view matching:(BOOL(^)(UIView *view))test
 {
