+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "guff",
+  "name": "Practica",
   "version": "0.0.1",
   "private": true,
   "scripts": {
